createUsser + encrypted refactorized and operative
