+++ helpers/generateJWT.js
@@ -3,8 +3,9 @@ const jwt = require('jsonwebtoken');
 const generateToken = (user) => {
     
     return new Promise ((resolve,reject) => {
+        
         const payload = {...user}
-        console.log('esto es payload',payload)
+
         jwt.sign(
             payload,
             process.env.JWT_SECRET_KEY,
